@@ -1,6 +1,6 @@
 export = ModelManager;
 /**
- * Manages the Concerto model files.
+ * Manages the Concerto model files in CTO format.
  *
  * The structure of {@link Resource}s (Assets, Transactions, Participants) is modelled
  * in a set of Concerto files. The contents of these files are managed
@@ -10,246 +10,24 @@ export = ModelManager;
  * Concerto applications load their Concerto files and then call the {@link ModelManager#addModelFile addModelFile}
  * method to register the Concerto file(s) with the ModelManager.
  *
- * Use the {@link Concerto} class to validate instances.
- *
  * @memberof module:concerto-core
  */
-declare class ModelManager {
+declare class ModelManager extends BaseModelManager {
     /**
      * Create the ModelManager.
      * @constructor
      * @param {object} [options] - Serializer options
      */
     constructor(options?: object);
-    modelFiles: {};
-    factory: Factory;
-    serializer: Serializer;
-    decoratorFactories: any[];
     /**
-     * Returns true
-     * @returns {boolean} true
-     */
-    isModelManager(): boolean;
-    /**
-     * Adds root types
-     * @private
-     */
-    private addRootModel;
-    /**
-     * Visitor design pattern
-     * @param {Object} visitor - the visitor
-     * @param {Object} parameters  - the parameter
-     * @return {Object} the result of visiting or null
-     */
-    accept(visitor: any, parameters: any): any;
-    /**
-     * Validates a Concerto file (as a string) to the ModelManager.
-     * Concerto files have a single namespace.
-     *
-     * Note that if there are dependencies between multiple files the files
-     * must be added in dependency order, or the addModelFiles method can be
-     * used to add a set of files irrespective of dependencies.
-     * @param {string|ModelFile} modelFile - The Concerto file as a string
-     * @param {string} [fileName] - a file name to associate with the model file
-     * @throws {IllegalModelException}
-     */
-    validateModelFile(modelFile: string | ModelFile, fileName?: string): void;
-    /**
-     * Throws an error with details about the existing namespace.
-     * @param {ModelFile} modelFile The model file that is trying to declare an existing namespace
-     * @private
-     */
-    private _throwAlreadyExists;
-    /**
-     * Adds a Concerto file (as a string) to the ModelManager.
-     * Concerto files have a single namespace. If a Concerto file with the
-     * same namespace has already been added to the ModelManager then it
-     * will be replaced.
-     * Note that if there are dependencies between multiple files the files
-     * must be added in dependency order, or the addModelFiles method can be
-     * used to add a set of files irrespective of dependencies.
-     * @param {string|ModelFile} modelFile - Model as a string or object
-     * @param {string} fileName - an optional file name to associate with the model file
+     * Adds a model in CTO format to the ModelManager.
+     * This is a convenience function equivalent to `addModel` but useful since it avoids having to copy the input CTO.
+     * @param {string} cto - a cto string
+     * @param {string} [fileName] - an optional file name to associate with the model file
      * @param {boolean} [disableValidation] - If true then the model files are not validated
      * @throws {IllegalModelException}
      * @return {Object} The newly added model file (internal).
      */
-    addModelFile(modelFile: string | ModelFile, fileName: string, disableValidation?: boolean): any;
-    /**
-     * Updates a Concerto file (as a string) on the ModelManager.
-     * Concerto files have a single namespace. If a Concerto file with the
-     * same namespace has already been added to the ModelManager then it
-     * will be replaced.
-     * @param {string|ModelFile} modelFile - Model as a string or object
-     * @param {string} [fileName] - a file name to associate with the model file
-     * @param {boolean} [disableValidation] - If true then the model files are not validated
-     * @throws {IllegalModelException}
-     * @returns {Object} The newly added model file (internal).
-     */
-    updateModelFile(modelFile: string | ModelFile, fileName?: string, disableValidation?: boolean): any;
-    /**
-     * Remove the Concerto file for a given namespace
-     * @param {string} namespace - The namespace of the model file to delete.
-     */
-    deleteModelFile(namespace: string): void;
-    /**
-     * Add a set of Concerto files to the model manager.
-     * @param {string[]|ModelFile[]} modelFiles - An array of models as strings or ModelFile objects.
-     * @param {string[]} [fileNames] - A array of file names to associate with the model files
-     * @param {boolean} [disableValidation] - If true then the model files are not validated
-     * @returns {Object[]} The newly added model files (internal).
-     */
-    addModelFiles(modelFiles: string[] | ModelFile[], fileNames?: string[], disableValidation?: boolean): any[];
-    /**
-     * Validates all models files in this model manager
-     */
-    validateModelFiles(): void;
-    /**
-     * Downloads all ModelFiles that are external dependencies and adds or
-     * updates them in this ModelManager.
-     * @param {Object} [options] - Options object passed to ModelFileLoaders
-     * @param {FileDownloader} [fileDownloader] - an optional FileDownloader
-     * @throws {IllegalModelException} if the models fail validation
-     * @return {Promise} a promise when the download and update operation is completed.
-     */
-    updateExternalModels(options?: any, fileDownloader?: typeof import("@accordproject/concerto-util/types/lib/filedownloader")): Promise<any>;
-    /**
-     * Write all models in this model manager to the specified path in the file system
-     *
-     * @param {string} path to a local directory
-     * @param {Object} [options] - Options object
-     * @param {boolean} options.includeExternalModels -
-     *  If true, external models are written to the file system. Defaults to true
-     */
-    writeModelsToFileSystem(path: string, options?: {
-        includeExternalModels: boolean;
-    }): void;
-    /**
-     * Get the array of model file instances
-     * @param {Boolean} [includeConcertoNamespace] - whether to include the concerto namespace
-     * (default to false)
-     * @return {ModelFile[]} The ModelFiles registered
-     * @private
-     */
-    private getModelFiles;
-    /**
-     * Gets all the Concerto models
-     * @param {Object} [options] - Options object
-     * @param {boolean} options.includeExternalModels -
-     *  If true, external models are written to the file system. Defaults to true
-     * @return {Array<{name:string, content:string}>} the name and content of each CTO file
-     */
-    getModels(options?: {
-        includeExternalModels: boolean;
-    }): Array<{
-        name: string;
-        content: string;
-    }>;
-    /**
-     * Check that the type is valid and returns the FQN of the type.
-     * @param {string} context - error reporting context
-     * @param {string} type - fully qualified type name
-     * @return {string} - the resolved type name (fully qualified)
-     * @throws {IllegalModelException} - if the type is not defined
-     * @private
-     */
-    private resolveType;
-    /**
-     * Remove all registered Concerto files
-     */
-    clearModelFiles(): void;
-    /**
-     * Get the ModelFile associated with a namespace
-     *
-     * @param {string} namespace - the namespace containing the ModelFile
-     * @return {ModelFile} registered ModelFile for the namespace or null
-     */
-    getModelFile(namespace: string): ModelFile;
-    /**
-     * Get the ModelFile associated with a file name
-     *
-     * @param {string} fileName - the fileName associated with the ModelFile
-     * @return {ModelFile} registered ModelFile for the namespace or null
-     * @private
-     */
-    private getModelFileByFileName;
-    /**
-     * Get the namespaces registered with the ModelManager.
-     * @return {string[]} namespaces - the namespaces that have been registered.
-     */
-    getNamespaces(): string[];
-    /**
-     * Look up a type in all registered namespaces.
-     *
-     * @param {string} qualifiedName - fully qualified type name.
-     * @return {ClassDeclaration} - the class declaration for the specified type.
-     * @throws {TypeNotFoundException} - if the type cannot be found or is a primitive type.
-     */
-    getType(qualifiedName: string): ClassDeclaration;
-    /**
-     * Get the AssetDeclarations defined in this model manager
-     * @return {AssetDeclaration[]} the AssetDeclarations defined in the model manager
-     */
-    getAssetDeclarations(): AssetDeclaration[];
-    /**
-     * Get the TransactionDeclarations defined in this model manager
-     * @return {TransactionDeclaration[]} the TransactionDeclarations defined in the model manager
-     */
-    getTransactionDeclarations(): TransactionDeclaration[];
-    /**
-     * Get the EventDeclarations defined in this model manager
-     * @return {EventDeclaration[]} the EventDeclaration defined in the model manager
-     */
-    getEventDeclarations(): EventDeclaration[];
-    /**
-     * Get the ParticipantDeclarations defined in this model manager
-     * @return {ParticipantDeclaration[]} the ParticipantDeclaration defined in the model manager
-     */
-    getParticipantDeclarations(): ParticipantDeclaration[];
-    /**
-     * Get the EnumDeclarations defined in this model manager
-     * @return {EnumDeclaration[]} the EnumDeclaration defined in the model manager
-     */
-    getEnumDeclarations(): EnumDeclaration[];
-    /**
-     * Get the Concepts defined in this model manager
-     * @return {ConceptDeclaration[]} the ConceptDeclaration defined in the model manager
-     */
-    getConceptDeclarations(): ConceptDeclaration[];
-    /**
-     * Get a factory for creating new instances of types defined in this model manager.
-     * @return {Factory} A factory for creating new instances of types defined in this model manager.
-     */
-    getFactory(): Factory;
-    /**
-     * Get a serializer for serializing instances of types defined in this model manager.
-     * @return {Serializer} A serializer for serializing instances of types defined in this model manager.
-     */
-    getSerializer(): Serializer;
-    /**
-     * Get the decorator factories for this model manager.
-     * @return {DecoratorFactory[]} The decorator factories for this model manager.
-     */
-    getDecoratorFactories(): DecoratorFactory[];
-    /**
-     * Add a decorator factory to this model manager.
-     * @param {DecoratorFactory} factory The decorator factory to add to this model manager.
-     */
-    addDecoratorFactory(factory: DecoratorFactory): void;
-    /**
-     * Checks if this fully qualified type name is derived from another.
-     * @param {string} fqt1 The fully qualified type name to check.
-     * @param {string} fqt2 The fully qualified type name it is may be derived from.
-     * @returns {boolean} True if this instance is an instance of the specified fully
-     * qualified type name, false otherwise.
-     */
-    derivesFrom(fqt1: string, fqt2: string): boolean;
-    /**
-     * Get the full metamodel for a modelmanager
-     * @returns {*} the metamodel
-     */
-    getMetaModel(): any;
+    addCTOModel(cto: string, fileName?: string, disableValidation?: boolean): any;
 }
-import Factory = require("./factory");
-import Serializer = require("./serializer");
-import ModelFile = require("./introspect/modelfile");
+import BaseModelManager = require("./basemodelmanager");
