@@ -82,7 +82,10 @@ class Identifiable extends Typed {
      * @return {string} the fully qualified identifier of this instance
      */
     getFullyQualifiedIdentifier() {
-        return this.getFullyQualifiedType() + '#' + (this.$identifier || 'null');
+        if (this.$identifier) {
+            return this.getFullyQualifiedType() + '#' + this.$identifier;
+        }
+        return this.getFullyQualifiedType();
     }
 
     /**
