@@ -52,7 +52,7 @@ describe('Introspector', () => {
             let modelBase = fs.readFileSync('./test/data/model/model-base.cto', 'utf8');
             modelBase.should.not.be.null;
 
-            modelManager.addModelFile(modelBase, 'model-base.cto');
+            modelManager.addCTOModel(modelBase, 'model-base.cto');
             const introspector = new Introspector(modelManager);
             let classDecl = introspector.getClassDeclarations();
             classDecl.length.should.equal(44);
@@ -70,7 +70,7 @@ describe('Introspector', () => {
             let modelBase = fs.readFileSync('./test/data/model/model-base.cto', 'utf8');
             modelBase.should.not.be.null;
 
-            modelManager.addModelFile(modelBase, 'model-base.cto');
+            modelManager.addCTOModel(modelBase, 'model-base.cto');
             const introspector = new Introspector(modelManager);
             introspector.getClassDeclaration('org.acme.base.Person').should.not.be.null;
         });
