@@ -1,33 +1,33 @@
-export var SecurityException: typeof import("./lib/securityexception");
-export var IllegalModelException: typeof import("./lib/introspect/illegalmodelexception");
-export var TypeNotFoundException: typeof import("./lib/typenotfoundexception");
-export var Decorator: typeof import("./lib/introspect/decorator");
-export var DecoratorFactory: typeof import("./lib/introspect/decoratorfactory");
-export var ClassDeclaration: typeof import("./lib/introspect/classdeclaration");
-export var IdentifiedDeclaration: typeof import("./lib/introspect/identifieddeclaration");
-export var AssetDeclaration: typeof import("./lib/introspect/assetdeclaration");
-export var ConceptDeclaration: typeof import("./lib/introspect/conceptdeclaration");
-export var EnumValueDeclaration: typeof import("./lib/introspect/enumvaluedeclaration");
-export var EventDeclaration: typeof import("./lib/introspect/eventdeclaration");
-export var ParticipantDeclaration: typeof import("./lib/introspect/participantdeclaration");
-export var TransactionDeclaration: typeof import("./lib/introspect/transactiondeclaration");
-export var Property: typeof import("./lib/introspect/property");
-export var Field: typeof import("./lib/introspect/field");
-export var EnumDeclaration: typeof import("./lib/introspect/enumdeclaration");
-export var RelationshipDeclaration: typeof import("./lib/introspect/relationshipdeclaration");
-export var Typed: typeof import("./lib/model/typed");
-export var Identifiable: typeof import("./lib/model/identifiable");
-export var Relationship: typeof import("./lib/model/relationship");
-export var Resource: typeof import("./lib/model/resource");
-export var Factory: typeof import("./lib/factory");
-export var Globalize: typeof import("./lib/globalize");
-export var Introspector: typeof import("./lib/introspect/introspector");
-export var ModelFile: typeof import("./lib/introspect/modelfile");
-export var ModelManager: typeof import("./lib/modelmanager");
-export var Serializer: typeof import("./lib/serializer");
-export var ModelUtil: typeof import("./lib/modelutil");
-export var ModelLoader: typeof import("./lib/modelloader");
-export var DateTimeUtil: typeof import("./lib/datetimeutil");
-export var Concerto: typeof import("./lib/concerto");
-export var MetaModel: typeof import("./lib/introspect/metamodel");
-export var version: any;
+import SecurityException = require("./lib/securityexception");
+import IllegalModelException = require("./lib/introspect/illegalmodelexception");
+import TypeNotFoundException = require("./lib/typenotfoundexception");
+import Decorator = require("./lib/introspect/decorator");
+import DecoratorFactory = require("./lib/introspect/decoratorfactory");
+import ClassDeclaration = require("./lib/introspect/classdeclaration");
+import IdentifiedDeclaration = require("./lib/introspect/identifieddeclaration");
+import AssetDeclaration = require("./lib/introspect/assetdeclaration");
+import ConceptDeclaration = require("./lib/introspect/conceptdeclaration");
+import EnumValueDeclaration = require("./lib/introspect/enumvaluedeclaration");
+import EventDeclaration = require("./lib/introspect/eventdeclaration");
+import ParticipantDeclaration = require("./lib/introspect/participantdeclaration");
+import TransactionDeclaration = require("./lib/introspect/transactiondeclaration");
+import Property = require("./lib/introspect/property");
+import Field = require("./lib/introspect/field");
+import EnumDeclaration = require("./lib/introspect/enumdeclaration");
+import RelationshipDeclaration = require("./lib/introspect/relationshipdeclaration");
+import Typed = require("./lib/model/typed");
+import Identifiable = require("./lib/model/identifiable");
+import Relationship = require("./lib/model/relationship");
+import Resource = require("./lib/model/resource");
+import Factory = require("./lib/factory");
+import Globalize = require("./lib/globalize");
+import Introspector = require("./lib/introspect/introspector");
+import ModelFile = require("./lib/introspect/modelfile");
+import ModelManager = require("./lib/modelmanager");
+import Serializer = require("./lib/serializer");
+import ModelUtil = require("./lib/modelutil");
+import ModelLoader = require("./lib/modelloader");
+import DateTimeUtil = require("./lib/datetimeutil");
+import Concerto = require("./lib/concerto");
+import MetaModel = require("./lib/introspect/metamodel");
+export { SecurityException, IllegalModelException, TypeNotFoundException, Decorator, DecoratorFactory, ClassDeclaration, IdentifiedDeclaration, AssetDeclaration, ConceptDeclaration, EnumValueDeclaration, EventDeclaration, ParticipantDeclaration, TransactionDeclaration, Property, Field, EnumDeclaration, RelationshipDeclaration, Typed, Identifiable, Relationship, Resource, Factory, Globalize, Introspector, ModelFile, ModelManager, Serializer, ModelUtil, ModelLoader, DateTimeUtil, Concerto, MetaModel, version };
