@@ -19,6 +19,5 @@
  * @module concerto-tools
  */
 
-module.exports.FileWriter = require('./lib/filewriter');
-module.exports.CodeGen = require('./lib/codegen/codegen.js');
+module.exports.CodeGen = require('./lib/codegen/codegen');
 module.exports.version = require('./package.json');
