@@ -17,7 +17,8 @@
 const fs = require('fs');
 const mkdirp = require('mkdirp');
 const path = require('path');
-const Writer = require('@accordproject/concerto-util').Writer;
+
+const Writer = require('./writer');
 
 /**
  * FileWriter creates text files under a directory tree. It can be used
